+++ electron-builder.yml
@@ -7,6 +7,7 @@ buildDependenciesFromSource: true
 
 extraResources:
   - './splashscreen.html'
+  - './updaterScreen.html'
   - 'resources/**'
   - './roboticsacademy.db'
 files:
