+++ src/main/index.ts
@@ -30,7 +30,7 @@ import {
   updateCommands,
   updateCommandUtils
 } from './db'
-import { createUpdaterWindow } from './appWindow'
+import { createSplashWindow, createUpdaterWindow, createWindow } from './appWindow'
 
 const isMac = process.platform === 'darwin'
 let mainWindow: BrowserWindow | null = null
@@ -43,112 +43,16 @@ autoUpdater.autoDownload = false
 // connected to database
 export const db: Database = dbInit()
 
-// splash screen
-function createSplashWindow(): BrowserWindow {
-  const win = new BrowserWindow({
-    width: 525,
-    height: 300,
-    frame: false,
-    transparent: true,
-    alwaysOnTop: true,
-    webPreferences: {
-      nodeIntegration: false,
-      webSecurity: true,
-      sandbox: false,
-      contextIsolation: true
-    }
-  })
-
-  const splashScreenSrc = app.isPackaged
-    ? join(process.resourcesPath, 'splashscreen.html')
-    : join(__dirname, './../../', 'splashscreen.html')
-
-  win.loadFile(splashScreenSrc)
-
-  win.webContents.on('did-finish-load', () => {
-    win.webContents
-      .executeJavaScript(
-        `
-        const version = document.getElementById('version')
-        version.textContent = '${appVersion}'
-        `
-      )
-      .then((result) => {
-        console.log('Executed in renderer:', result)
-      })
-      .catch(console.error)
-  })
-
-  return win
-}
-
-//  updater window
-
-// main window
-const createWindow = async (): Promise<BrowserWindow> => {
-  mainWindow = new BrowserWindow({
-    width: 1280,
-    height: 720,
-    minWidth: 1280,
-    minHeight: 720,
-    show: false,
-    frame: false,
-    alwaysOnTop: false,
-    transparent: false,
-    focusable: true,
-    icon: join(__dirname, './../../resources/icons/icon.png'),
-    autoHideMenuBar: true,
-    webPreferences: {
-      preload: join(__dirname, '../preload/index.js'),
-      nodeIntegration: false,
-      webSecurity: true,
-      sandbox: false,
-      contextIsolation: true
-    }
-  })
-
-  //
-  mainWindow.webContents.session.webRequest.onHeadersReceived((details, callback) => {
-    if (details.responseHeaders === undefined) return
-    callback({
-      responseHeaders: Object.fromEntries(
-        Object.entries(details.responseHeaders).filter(
-          (header) => !/x-frame-options/i.test(header[0])
-        )
-      )
-    })
-  })
-  mainWindow.webContents.setWindowOpenHandler((details) => {
-    shell.openExternal(details.url)
-    return { action: 'deny' }
-  })
-
-  // Load the remote URL for development or the local html file for production.
-  if (is.dev && process.env['ELECTRON_RENDERER_URL']) {
-    mainWindow.loadURL(process.env['ELECTRON_RENDERER_URL'])
-  } else {
-    mainWindow.loadFile(join(__dirname, '../renderer/index.html'))
-  }
-
-  // window resize/close func
-  mainWindow.on('closed', (_e) => {
-    mainWindow = null
-  })
-  mainWindow.on('maximize', () => {})
-  mainWindow.on('unmaximize', () => {})
-  // Event listener for when the window is restored from minimized state
-  mainWindow.on('restore', () => {})
-
-  return mainWindow
-}
-
 app.whenReady().then(async () => {
   // disable http-cache
   app.commandLine.appendSwitch('disable-http-cache')
 
   // check update
-  autoUpdater.checkForUpdates()
-
+  try {
+    autoUpdater.checkForUpdates()
+  } catch (error) {
+    console.log(error)
+  }
   //Store data
   await insertCommandData(db)
   await insertCommandUtilsData(db)
@@ -224,6 +128,8 @@ app.whenReady().then(async () => {
   //* App Window Resize
   // minimize the window
   ipcMain.on('app_window:MINIMIZE', (_event) => {
+    if (mainWindow === null || mainWindow === undefined) return
+
     if (!mainWindow?.isMinimized()) {
       mainWindow?.minimize()
     }
@@ -401,24 +307,29 @@ app.whenReady().then(async () => {
   ipcMain.handle('updater:OPEN_LINK', (_event, url: string) => {
     try {
       shell.openExternal(url)
-    } catch (error) {}
+    } catch (error) {
+      console.log(error)
+    }
   })
   ipcMain.handle('updater:CLOSE_WINDOW', (_event) => {
     try {
-      if (updaterWindow) updaterWindow.destroy()
+      if (updaterWindow != null) updaterWindow.destroy()
     } catch (error) {}
   })
   //@ Disappering splash screen and show main screen after 3 seconds.
   try {
-    const splashScreen: BrowserWindow = createSplashWindow()
-    updaterWindow = createUpdaterWindow()
-    const mainScreen: BrowserWindow = await createWindow()
+    const splashScreen: BrowserWindow = createSplashWindow({ appVersion })
+    // const mainScreen: BrowserWindow = await createWindow()
 
-    updaterWindow.show()
-    mainScreen.once('ready-to-show', () => {
+    mainWindow = await createWindow()
+    // updaterWindow.show()
+
+    mainWindow?.once('ready-to-show', () => {
       setTimeout(
         () => {
-          // mainScreen.show()
+          // updaterWindow = createUpdaterWindow('2.0.1')
+          // updaterWindow.show()
+          mainWindow?.show()
           splashScreen.destroy()
         },
         app.isPackaged ? 3000 : 0
@@ -459,7 +370,8 @@ app.on('window-all-closed', (_e) => {
 
 // Auto Updater
 autoUpdater.on('update-available', (info) => {
-  console.log('====================================')
-  console.log('info ', info)
-  console.log('====================================')
+  updaterWindow = createUpdaterWindow(info.version)
+  setTimeout(() => {
+    updaterWindow?.show()
+  }, 10 * 1000)
 })
